sdk/python/requirements.txt
pyyaml
requests
python-crontab
